mll/turk/webservice/requirements.txt
aiohttp-cors
numpy
torch
sqlalchemy
pillow
scipy
aiohttp-devtools
